refactor: 패키지(package) 분리
- filter 클래스 및 연관 클래스 패키지 분리
  config -> interfaces/filter
