+++ src/main/java/io/hhplus/conbook/interfaces/filter/TokenAuthenticationFilter.java
@@ -1,7 +1,8 @@
-package io.hhplus.conbook.config;
+package io.hhplus.conbook.interfaces.filter;
 
 import io.hhplus.conbook.domain.token.AccessTokenInfo;
 import io.hhplus.conbook.domain.token.TokenManager;
+import io.hhplus.conbook.interfaces.api.ApiRoutes;
 import io.hhplus.conbook.interfaces.api.ErrorCode;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -27,8 +28,15 @@ public class TokenAuthenticationFilter extends OncePerRequestFilter {
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        String authorizationHeader = request.getHeader(JWT_AUTHORIZATION);
         String requestURI = request.getRequestURI();
+
+        if (requestURI.contains(ApiRoutes.BASE_TOKEN_API_PATH)) {
+            log.info("url: {}", requestURI);
+
+            filterChain.doFilter(request, response);
+            return;
+        }
+        String authorizationHeader = request.getHeader(JWT_AUTHORIZATION);
         log.info("token: {}, url: {}", authorizationHeader, requestURI);
 
         if (!StringUtils.hasText(authorizationHeader))
